fix: move update to utils
  - update.py
    * chore: drop auto_update options
    * fix: reloading bench modules post requirements setup (seemed
    unecessary and randomly crashed `bench start` process)
    * docs: added help for commands
    * chore: deprecated switch-to-master function

  - utils.py
    * sorted imports for readability
    * formatting and consistency
    * dropped unused setup_auto_update
    * dropped broken pre_upgrade function
